+++ app/Http/Controllers/PembayaranUangMakanController.php
@@ -5,28 +5,60 @@ namespace App\Http\Controllers;
 use App\Models\bulan;
 use Illuminate\Http\Request;
 use App\Models\dokumenUangMakan;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class PembayaranUangMakanController extends Controller
 {
     public function index($thn = null)
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
         if (!$thn) {
             $thn = date('Y');
         }
-        $kdsatker= 411792;
+        $kdsatker= auth()->user()->kdsatker;
         $data = dokumenUangMakan::uangMakan($kdsatker, $thn)->get();
-        $tahun = dokumenUangMakan::tahun(411792);
+        $tahun = dokumenUangMakan::tahun($kdsatker);
         return view('Pembayaran.Uang_makan.index',[
             'data'=>$data,
             'tahun'=>$tahun,
             'thn'=>$thn,
-            "pageTitle"=>"Dokumen Uang Lembur",
+            "pageTitle"=>"Uang Makan",
         ]);
     }
 
     public function create()
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
         return view('Pembayaran.Uang_makan.create',[
             'bulan'=>bulan::orderBy('bulan')->get()  
         ]);
@@ -34,6 +66,21 @@ class PembayaranUangMakanController extends Controller
 
     public function store(Request $request)
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
         $request->validate([
             'bulan'=>'required',
             'jmlpegawai'=>'required|numeric',
@@ -47,15 +94,33 @@ class PembayaranUangMakanController extends Controller
             'bulan'=>$request->bulan,
             'jmlpegawai'=>$request->jmlpegawai,
             'keterangan'=>$request->keterangan,
-            'kdsatker'=>411792,
+            'kdsatker'=>auth()->user()->kdsatker,
             'file'=>$path,
             'nmbulan'=>$nmbulan
         ]);
-        return Redirect('/pembayaran/uang-makan/index')->with('berhasil', 'data berhasil ditambahkan');
+        return Redirect('/belanja-51/uang-makan/index')->with('berhasil', 'data berhasil ditambahkan');
     }
 
     public function edit(dokumenUangMakan $dokumenUangMakan)
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
+        if ($dokumenUangMakan->kdsatker != auth()->user()->kdsatker) {
+            abort(403);
+        }
         if ($dokumenUangMakan->terkirim) {
             return abort(403);
         }
@@ -67,6 +132,24 @@ class PembayaranUangMakanController extends Controller
 
     public function update(Request $request, dokumenUangMakan $dokumenUangMakan)
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
+        if ($dokumenUangMakan->kdsatker != auth()->user()->kdsatker) {
+            abort(403);
+        }
         if ($dokumenUangMakan->terkirim) {
             return abort(403);
         }
@@ -103,27 +186,63 @@ class PembayaranUangMakanController extends Controller
                 'nmbulan'=>$nmbulan
             ]);
         }
-        return Redirect('/pembayaran/uang-makan/index')->with('berhasil', 'data berhasil di ubah');
+        return Redirect('/belanja-51/uang-makan/index')->with('berhasil', 'data berhasil di ubah');
     }
 
     public function delete(dokumenUangMakan $dokumenUangMakan)
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
+        if ($dokumenUangMakan->kdsatker != auth()->user()->kdsatker) {
+            abort(403);
+        }
         if ($dokumenUangMakan->terkirim) {
             return abort(403);
         }
         Storage::delete($dokumenUangMakan->file);
         $dokumenUangMakan->delete();
-        return Redirect('/pembayaran/uang-makan/index')->with('berhasil', 'data berhasil di hapus');
+        return Redirect('/belanja-51/uang-makan/index')->with('berhasil', 'data berhasil di hapus');
     }
 
     public function kirim(dokumenUangMakan $dokumenUangMakan)
     {
+        if (Auth::guard('web')->check()) {
+            $gate=['plt_admin_satker', 'opr_belanja_51'];
+            $gate2=['sys_admin'];
+        }else{
+            $gate=['admin_satker'];
+            $gate2=[];
+
+        }
+
+        if (! Gate::any($gate, auth()->user()->id)) {
+            if (! Gate::any($gate2, auth()->user()->id)) {
+                abort(403);
+            }
+            return Redirect('/belanja-51/dokumen-uang-makan');
+        }
+        if ($dokumenUangMakan->kdsatker != auth()->user()->kdsatker) {
+            abort(403);
+        }
         if ($dokumenUangMakan->terkirim) {
             return abort(403);
         }
         $dokumenUangMakan->update([
             'terkirim'=>true
         ]);
-        return Redirect('/pembayaran/uang-makan/index')->with('berhasil', 'data berhasil di kirim');
+        return Redirect('/belanja-51/uang-makan/index')->with('berhasil', 'data berhasil di kirim');
     }
 }
